requirements.txt
asyncio-nats-client
asyncio-nats-client[nkeys]
faker
requests
packaging
redis
async_property
